Fix Unwrap to support Go < 1.20
The code relies in the old way to unwrap things.
So it doesn't need an Unwrap method

This allows us to keep the feature without having to use Unwrap []error
and then need a //go:build tag.

Recent Go versions that support Unwrap []error still supports func
Is(error) bool and As(error) bool interfaces

Co-authored-by: WGH <wgh@torlan.ru>