Forbid unmarshaling non-scalars to TextUnmarshaler
This is a breaking change.

Previously an attempt to unmarshal a YAML mapping into a struct
implementing encoding.TextUnmarshaler would succeed, but likely do the wrong
thing.

Consider this example:

    type Foo struct {
        IP netip.Addr `yaml:"ip"`
    }

netip.Addr is a struct type implementing TextUnmarshaler, but
containing no exported fields. So unmarshaling {"ip": "127.0.0.1"} would
use TextUnmarshaler (as expected), but {"ip": {"whatever": 123}} would
silently do nothing.

With this commit, it will now result in a type error.

This matches the behavior of both encoding/json and encoding/json/v2.

The documentation of the latter states:

> If the value type implements encoding.TextUnmarshaler, then the input
> is decoded as a JSON string and the UnmarshalText method is called with
> the decoded string value. This fails with a SemanticError if the input
> is not a JSON string.

encoding/json documentation is somewhat ambiguous on this, but it also
behaves the same way (https://go.dev/play/p/z9okIXkUzvQ).