Change error message about decoding into Go map
YAML allows sequences and mappings as a key for a mapping.
Go does not support slices or maps as a map key; and the default
decoder returns an error "yaml: invalid map key %#v". but i feel
that this error message is misleading and may cause the user to
think that it's an error in the document.

i updated the error message to be indicate this situation and ask
the user to try `yaml.Node`