Configure golangci-lint to report all errors
By default, golangci-lint reports only:
- 3 errors per linter.
- 50 errors in total.
- only one error per line (the first one encountered).

These settings are good default values when enabling golangci-lint to a
project with a lot of code and issues to fix.

But it comes with a drawback, when someone updates code, only a limited
number of errors will be reported, letting user think the code is OK.

I faced many times when people are pushing code, see errors reported by
the CI in GitHub Actions, and then fix them, but after pushing again,
new errors are reported, which were already present in the code before,
but were not reported because of the limits.

This creates frustrations, and people may feel that the linting process
is not effective.

Here we configure golangci-lint to report all errors by adjusting the
settings in the .golangci.yaml file.