Adopt indent and knownFields options
This commit builds opon the previous one to adopt the new `Options`
framework for two existing parameters: `indent` and `knownFields` which
are currently controlled respectively through the `Encoder.SetIndent()`
and `Decoder.KnownFields()` methods.

With this commit, they can now additionally be set via `OptionIndent()`
and `OptionKnownFields()` which return `Options` that can be passed to
any public method and will get propagated down to the relevant place.

Signed-off-by: Hadrien Patte <hadrien.patte@protonmail.com>