add Node.DecodeWithOptions
This PR adds a `DecodeWithOptions()` method to `Node` which allows
strict unmarshaling. Without this, it is impossible to require strict
unmarshaling from inside a custom `UnmarshalYAML(node *yaml.Node)`
handler. This is important for many applications where we don't want a
typo in a configuration file to be silently ignored.

See https://github.com/go-yaml/yaml/issues/460 for more info. Note
that this is functionally a regression in v3/v4; in v2 the decoder
flags are automatically used by custom unmarshalers.

This is an adaptation of https://github.com/go-yaml/yaml/pull/691